+++ Sources/DeckKit/_Deprecated/StackedDeck+Deprecated.swift
@@ -41,3 +41,43 @@ extension StackedDeck {
         )
     }
 }
+
+
+@available(*, deprecated, message: "Use the itemDisplayCount initializer instead.")
+extension StackedDeckConfiguration {
+
+    /**
+     Create a stacked deck configuration.
+
+     - Parameters:
+       - direction: The visual direction of the stack, by default `.up`.
+       - cardDisplayCount: The max number of cards to display, by default `10`.
+       - alwaysShowLastCard: Whether or not to show the last card for visual stability, by default `true`.
+       - scaleOffset: The percentual factor to shrink cards for each step down the stack, by default `0.02`.
+       - verticalOffset: The vertical offset to apply to cards for each step down the stack, by default `10`.
+       - dragRotationFactor: The offset factor with which to rotate a card when it's panned, by default `0.05`.
+       - horizontalDragThreshold: The number of points a card must be panned to be moved to the bottom of the deck, by default `100`.
+       - verticalDragThreshold: The number of points a card must be panned to be moved to the bottom of the deck, by default `250`.
+     */
+    public init(
+        direction: Direction = .up,
+        cardDisplayCount: Int,
+        alwaysShowLastCard: Bool = true,
+        scaleOffset: Double = 0.02,
+        verticalOffset: Double = 10,
+        dragRotationFactor: Double = 0.01,
+        horizontalDragThreshold: Double = 100,
+        verticalDragThreshold: Double = 250
+    ) {
+        self.init(
+            direction: direction,
+            itemDisplayCount: cardDisplayCount,
+            alwaysShowLastCard: alwaysShowLastCard,
+            scaleOffset: scaleOffset,
+            verticalOffset: verticalOffset,
+            dragRotationFactor: dragRotationFactor,
+            horizontalDragThreshold: horizontalDragThreshold,
+            verticalDragThreshold: verticalDragThreshold
+        )
+    }
+}